move tables to new package
