+++ server/src/main/kotlin/ir/armor/tachidesk/impl/Category.kt
@@ -8,9 +8,9 @@ package ir.armor.tachidesk.impl
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import ir.armor.tachidesk.impl.CategoryManga.removeMangaFromCategory
-import ir.armor.tachidesk.model.database.CategoryMangaTable
-import ir.armor.tachidesk.model.database.CategoryTable
-import ir.armor.tachidesk.model.database.toDataClass
+import ir.armor.tachidesk.model.database.table.CategoryMangaTable
+import ir.armor.tachidesk.model.database.table.CategoryTable
+import ir.armor.tachidesk.model.database.table.toDataClass
 import ir.armor.tachidesk.model.dataclass.CategoryDataClass
 import org.jetbrains.exposed.sql.SortOrder
 import org.jetbrains.exposed.sql.deleteWhere
