+++ server/src/main/kotlin/ir/armor/tachidesk/impl/Chapter.kt
@@ -12,9 +12,9 @@ import eu.kanade.tachiyomi.source.model.SManga
 import ir.armor.tachidesk.impl.Manga.getManga
 import ir.armor.tachidesk.impl.util.GetHttpSource.getHttpSource
 import ir.armor.tachidesk.impl.util.awaitSingle
-import ir.armor.tachidesk.model.database.ChapterTable
-import ir.armor.tachidesk.model.database.MangaTable
-import ir.armor.tachidesk.model.database.PageTable
+import ir.armor.tachidesk.model.database.table.ChapterTable
+import ir.armor.tachidesk.model.database.table.MangaTable
+import ir.armor.tachidesk.model.database.table.PageTable
 import ir.armor.tachidesk.model.dataclass.ChapterDataClass
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.insert
