+++ server/src/main/kotlin/ir/armor/tachidesk/impl/Library.kt
@@ -8,9 +8,9 @@ package ir.armor.tachidesk.impl
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import ir.armor.tachidesk.impl.Manga.getManga
-import ir.armor.tachidesk.model.database.CategoryMangaTable
-import ir.armor.tachidesk.model.database.MangaTable
-import ir.armor.tachidesk.model.database.toDataClass
+import ir.armor.tachidesk.model.database.table.CategoryMangaTable
+import ir.armor.tachidesk.model.database.table.MangaTable
+import ir.armor.tachidesk.model.database.table.toDataClass
 import ir.armor.tachidesk.model.dataclass.MangaDataClass
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.deleteWhere
