+++ server/src/main/kotlin/ir/armor/tachidesk/impl/MangaList.kt
@@ -10,8 +10,8 @@ package ir.armor.tachidesk.impl
 import eu.kanade.tachiyomi.source.model.MangasPage
 import ir.armor.tachidesk.impl.util.GetHttpSource.getHttpSource
 import ir.armor.tachidesk.impl.util.awaitSingle
-import ir.armor.tachidesk.model.database.MangaStatus
-import ir.armor.tachidesk.model.database.MangaTable
+import ir.armor.tachidesk.model.database.table.MangaStatus
+import ir.armor.tachidesk.model.database.table.MangaTable
 import ir.armor.tachidesk.model.dataclass.MangaDataClass
 import ir.armor.tachidesk.model.dataclass.PagedMangaListDataClass
 import org.jetbrains.exposed.sql.insertAndGetId
