+++ server/src/main/kotlin/ir/armor/tachidesk/impl/backup/legacy/LegacyBackupImport.kt
@@ -22,7 +22,7 @@ import ir.armor.tachidesk.impl.backup.models.Track
 import ir.armor.tachidesk.impl.backup.models.TrackImpl
 import ir.armor.tachidesk.impl.util.GetHttpSource.getHttpSource
 import ir.armor.tachidesk.impl.util.awaitSingle
-import ir.armor.tachidesk.model.database.MangaTable
+import ir.armor.tachidesk.model.database.table.MangaTable
 import mu.KotlinLogging
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.insert
