+++ server/src/main/kotlin/ir/armor/tachidesk/model/database/table/CategoryTable.kt
@@ -1,4 +1,4 @@
-package ir.armor.tachidesk.model.database
+package ir.armor.tachidesk.model.database.table
 
 /*
  * Copyright (C) Contributors to the Suwayomi project
@@ -19,7 +19,7 @@ object CategoryTable : IntIdTable() {
 
 fun CategoryTable.toDataClass(categoryEntry: ResultRow) = CategoryDataClass(
     categoryEntry[CategoryTable.id].value,
-    categoryEntry[CategoryTable.order],
-    categoryEntry[CategoryTable.name],
-    categoryEntry[CategoryTable.isLanding],
+    categoryEntry[order],
+    categoryEntry[name],
+    categoryEntry[isLanding],
 )
